+++ docs/modules/spark-k8s/examples/getting_started/test_getting_started_stackablectl.sh
@@ -3,4 +3,3 @@ set -euo pipefail
 
 cd "$(dirname "$0")"
 ./getting_started.sh stackablectl
-