+++ rust/crd/src/history.rs
@@ -1,8 +1,11 @@
+use crate::s3logdir::S3LogDir;
+use crate::tlscerts;
 use crate::{affinity::history_affinity, constants::*};
 
 use product_config::{types::PropertyNameKind, ProductConfigManager};
 use serde::{Deserialize, Serialize};
 use snafu::{OptionExt, ResultExt, Snafu};
+use stackable_operator::k8s_openapi::api::core::v1::EnvVar;
 use stackable_operator::role_utils::RoleGroup;
 use stackable_operator::{
     commons::{
@@ -37,11 +40,11 @@ use strum::{Display, EnumIter};
 pub enum Error {
     #[snafu(display("failed to transform configs"))]
     ProductConfigTransform {
-        source: stackable_operator::product_config_utils::ConfigError,
+        source: stackable_operator::product_config_utils::Error,
     },
     #[snafu(display("invalid product config"))]
     InvalidProductConfig {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::product_config_utils::Error,
     },
     #[snafu(display("fragment validation failure"))]
     FragmentValidationFailure { source: ValidationError },
@@ -232,6 +235,96 @@ impl SparkHistoryServer {
         )
         .context(InvalidProductConfigSnafu)
     }
+
+    pub fn merged_env(
+        &self,
+        s3logdir: &S3LogDir,
+        role_group_env_overrides: HashMap<String, String>,
+    ) -> Vec<EnvVar> {
+        // Maps env var name to env var object. This allows env_overrides to work
+        // as expected (i.e. users can override the env var value).
+        let mut vars: BTreeMap<String, EnvVar> = BTreeMap::new();
+        let role_env_overrides = &self.role().config.env_overrides;
+
+        // This env var prevents the history server from detaching itself from the
+        // start script because this leads to the Pod terminating immediately.
+        vars.insert(
+            "SPARK_NO_DAEMONIZE".to_string(),
+            EnvVar {
+                name: "SPARK_NO_DAEMONIZE".to_string(),
+                value: Some("true".into()),
+                value_from: None,
+            },
+        );
+        vars.insert(
+            "SPARK_DAEMON_CLASSPATH".to_string(),
+            EnvVar {
+                name: "SPARK_DAEMON_CLASSPATH".to_string(),
+                value: Some("/stackable/spark/extra-jars/*".into()),
+                value_from: None,
+            },
+        );
+
+        let mut history_opts = vec![
+        format!("-Dlog4j.configurationFile={VOLUME_MOUNT_PATH_LOG_CONFIG}/{LOG4J2_CONFIG_FILE}"),
+        format!(
+            "-Djava.security.properties={VOLUME_MOUNT_PATH_CONFIG}/{JVM_SECURITY_PROPERTIES_FILE}"
+        ),
+        format!("-javaagent:/stackable/jmx/jmx_prometheus_javaagent.jar={METRICS_PORT}:/stackable/jmx/config.yaml")
+    ];
+
+        // if TLS is enabled build truststore
+        if tlscerts::tls_secret_name(&s3logdir.bucket.connection).is_some() {
+            history_opts.extend(vec![
+                format!("-Djavax.net.ssl.trustStore={STACKABLE_TRUST_STORE}/truststore.p12"),
+                format!("-Djavax.net.ssl.trustStorePassword={STACKABLE_TLS_STORE_PASSWORD}"),
+                format!("-Djavax.net.ssl.trustStoreType=pkcs12"),
+            ]);
+        }
+
+        vars.insert(
+            "SPARK_HISTORY_OPTS".to_string(),
+            EnvVar {
+                name: "SPARK_HISTORY_OPTS".to_string(),
+                value: Some(history_opts.join(" ")),
+                value_from: None,
+            },
+        );
+
+        // apply the role overrides
+        let mut role_envs = role_env_overrides.iter().map(|(env_name, env_value)| {
+            (
+                env_name.clone(),
+                EnvVar {
+                    name: env_name.clone(),
+                    value: Some(env_value.to_owned()),
+                    value_from: None,
+                },
+            )
+        });
+
+        vars.extend(&mut role_envs);
+
+        // apply the role-group overrides
+        let mut role_group_envs =
+            role_group_env_overrides
+                .into_iter()
+                .map(|(env_name, env_value)| {
+                    (
+                        env_name.clone(),
+                        EnvVar {
+                            name: env_name.clone(),
+                            value: Some(env_value),
+                            value_from: None,
+                        },
+                    )
+                });
+
+        vars.extend(&mut role_group_envs);
+
+        // convert to Vec
+        vars.into_values().collect()
+    }
 }
 
 #[derive(Clone, Debug, Deserialize, JsonSchema, Serialize, Display)]
@@ -339,7 +432,7 @@ impl Configuration for HistoryConfigFragment {
         &self,
         _resource: &Self::Configurable,
         _role_name: &str,
-    ) -> stackable_operator::product_config_utils::ConfigResult<BTreeMap<String, Option<String>>>
+    ) -> Result<BTreeMap<String, Option<String>>, stackable_operator::product_config_utils::Error>
     {
         Ok(BTreeMap::new())
     }
@@ -348,7 +441,7 @@ impl Configuration for HistoryConfigFragment {
         &self,
         _resource: &Self::Configurable,
         _role_name: &str,
-    ) -> stackable_operator::product_config_utils::ConfigResult<BTreeMap<String, Option<String>>>
+    ) -> Result<BTreeMap<String, Option<String>>, stackable_operator::product_config_utils::Error>
     {
         Ok(BTreeMap::new())
     }
@@ -358,8 +451,79 @@ impl Configuration for HistoryConfigFragment {
         _resource: &Self::Configurable,
         _role_name: &str,
         _file: &str,
-    ) -> stackable_operator::product_config_utils::ConfigResult<BTreeMap<String, Option<String>>>
+    ) -> Result<BTreeMap<String, Option<String>>, stackable_operator::product_config_utils::Error>
     {
         Ok(BTreeMap::new())
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+    use indoc::indoc;
+    use stackable_operator::commons::s3::InlinedS3BucketSpec;
+
+    #[test]
+    pub fn test_env_overrides() {
+        let input = indoc! {r#"
+        ---
+        apiVersion: spark.stackable.tech/v1alpha1
+        kind: SparkHistoryServer
+        metadata:
+          name: spark-history
+        spec:
+          image:
+            productVersion: 3.5.1
+          logFileDirectory:
+            s3:
+              prefix: eventlogs/
+              bucket:
+                reference: spark-history-s3-bucket
+          nodes:
+            envOverrides:
+              TEST_SPARK_HIST_VAR: ROLE
+            roleGroups:
+              default:
+                replicas: 1
+                config:
+                  cleaner: true
+                envOverrides:
+                  TEST_SPARK_HIST_VAR: ROLEGROUP
+        "#};
+
+        let deserializer = serde_yaml::Deserializer::from_str(input);
+        let history: SparkHistoryServer =
+            serde_yaml::with::singleton_map_recursive::deserialize(deserializer).unwrap();
+
+        let s3_log_dir: S3LogDir = S3LogDir {
+            bucket: InlinedS3BucketSpec {
+                bucket_name: None,
+                connection: None,
+            },
+            prefix: "prefix".to_string(),
+        };
+
+        let merged_env = history.merged_env(
+            &s3_log_dir,
+            history
+                .spec
+                .nodes
+                .role_groups
+                .get("default")
+                .unwrap()
+                .config
+                .env_overrides
+                .clone(),
+        );
+
+        let env_map: BTreeMap<&str, Option<String>> = merged_env
+            .iter()
+            .map(|env_var| (env_var.name.as_str(), env_var.value.clone()))
+            .collect();
+
+        assert_eq!(
+            Some(&Some("ROLEGROUP".to_string())),
+            env_map.get("TEST_SPARK_HIST_VAR")
+        );
     }
 }