+++ rust/crd/src/lib.rs
@@ -14,14 +14,16 @@ use product_config::{types::PropertyNameKind, ProductConfigManager};
 use s3logdir::S3LogDir;
 use serde::{Deserialize, Serialize};
 use snafu::{OptionExt, ResultExt, Snafu};
-use stackable_operator::builder::SecretFormat;
 use stackable_operator::product_config_utils::{
     transform_all_roles_to_config, validate_all_roles_and_groups_config,
     ValidatedRoleConfigByPropertyKind,
 };
 use stackable_operator::role_utils::EmptyRoleConfig;
 use stackable_operator::{
-    builder::{SecretOperatorVolumeSourceBuilder, VolumeBuilder},
+    builder::pod::volume::{
+        SecretFormat, SecretOperatorVolumeSourceBuilder, SecretOperatorVolumeSourceBuilderError,
+        VolumeBuilder,
+    },
     commons::{
         product_image_selection::{ProductImage, ResolvedProductImage},
         resources::{CpuLimits, MemoryLimits, Resources},
@@ -38,6 +40,7 @@ use stackable_operator::{
     product_logging,
     role_utils::{CommonConfiguration, Role, RoleGroup},
     schemars::{self, JsonSchema},
+    utils::crds::raw_object_list_schema,
 };
 use std::{
     cmp::max,
@@ -63,7 +66,7 @@ pub enum Error {
 
     #[snafu(display("failed to convert java heap config to unit [{unit}]"))]
     FailedToConvertJavaHeap {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::memory::Error,
         unit: String,
     },
 
@@ -75,17 +78,17 @@ pub enum Error {
 
     #[snafu(display("failed to transform configs"))]
     ProductConfigTransform {
-        source: stackable_operator::product_config_utils::ConfigError,
+        source: stackable_operator::product_config_utils::Error,
     },
 
     #[snafu(display("invalid product config"))]
     InvalidProductConfig {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::product_config_utils::Error,
     },
 
     #[snafu(display("failed to build TLS certificate SecretClass Volume"))]
     TlsCertSecretClassVolumeBuild {
-        source: stackable_operator::builder::SecretOperatorVolumeSourceBuilderError,
+        source: SecretOperatorVolumeSourceBuilderError,
     },
 
     #[snafu(display("failed to build S3 credentials Volume"))]
@@ -153,6 +156,8 @@ pub struct SparkApplicationSpec {
 
     /// The job builds a spark-submit command, complete with arguments and referenced dependencies
     /// such as templates, and passes it on to Spark.
+    /// The reason this property uses its own type (SubmitConfigFragment) is because logging is not
+    /// supported for spark-submit processes.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub job: Option<CommonConfiguration<SubmitConfigFragment>>,
 
@@ -163,6 +168,8 @@ pub struct SparkApplicationSpec {
 
     /// The executor role specifies the configuration that, together with the driver pod template, is used by
     /// Spark to create the executor pods.
+    /// This is RoleGroup instead of plain CommonConfiguration because it needs to allows for the number of replicas.
+    /// to be specified.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub executor: Option<RoleGroup<RoleConfigFragment>>,
 
@@ -187,6 +194,7 @@ pub struct SparkApplicationSpec {
 
     /// A list of volumes that can be made available to the job, driver or executors via their volume mounts.
     #[serde(default)]
+    #[schemars(schema_with = "raw_object_list_schema")]
     pub volumes: Vec<Volume>,
 
     /// A list of environment variables that will be set in the job pod and the driver and executor
@@ -343,12 +351,23 @@ impl SparkApplication {
         Ok(result)
     }
 
+    /// Return the volume mounts for the spark-submit pod.
+    ///
+    /// These volume mounts are assembled from:
+    /// * two pod template CMs for the driver and executors
+    /// * volume mounts for accessing applications stored in S3 buckets
+    /// * S3 credentials
+    /// * S3 verification certificates
+    /// * python packages (razvan: this was also a mistake since these packages are not used here.)
+    /// * volume mounts additional java packages
+    /// * finally user specified volume maps in `spec.job`.
+    ///
     pub fn spark_job_volume_mounts(
         &self,
         s3conn: &Option<S3ConnectionSpec>,
         s3logdir: &Option<S3LogDir>,
     ) -> Vec<VolumeMount> {
-        let volume_mounts = vec![
+        let mut tmpl_mounts = vec![
             VolumeMount {
                 name: VOLUME_MOUNT_NAME_DRIVER_POD_TEMPLATES.into(),
                 mount_path: VOLUME_MOUNT_PATH_DRIVER_POD_TEMPLATES.into(),
@@ -360,7 +379,25 @@ impl SparkApplication {
                 ..VolumeMount::default()
             },
         ];
-        self.add_common_volume_mounts(volume_mounts, s3conn, s3logdir, false)
+
+        tmpl_mounts = self.add_common_volume_mounts(tmpl_mounts, s3conn, s3logdir, false);
+
+        if let Some(CommonConfiguration {
+            config:
+                SubmitConfigFragment {
+                    volume_mounts:
+                        Some(VolumeMounts {
+                            volume_mounts: job_vm,
+                        }),
+                    ..
+                },
+            ..
+        }) = &self.spec.job
+        {
+            tmpl_mounts.extend(job_vm.clone());
+        }
+
+        tmpl_mounts
     }
 
     fn add_common_volume_mounts(
@@ -690,6 +727,37 @@ impl SparkApplication {
         }
     }
 
+    pub fn merged_env(&self, role: SparkApplicationRole, env: &[EnvVar]) -> Vec<EnvVar> {
+        // Use a BTreeMap internally to enable replacement of existing keys
+        let mut env: BTreeMap<&String, EnvVar> = env
+            .iter()
+            .map(|env_var| (&env_var.name, env_var.clone()))
+            .collect();
+
+        // Merge the role-specific envOverrides on top
+        let role_envs = match role {
+            SparkApplicationRole::Submit => self.spec.job.as_ref().map(|j| &j.env_overrides),
+            SparkApplicationRole::Driver => self.spec.driver.as_ref().map(|d| &d.env_overrides),
+            SparkApplicationRole::Executor => {
+                self.spec.executor.as_ref().map(|e| &e.config.env_overrides)
+            }
+        };
+        if let Some(role_envs) = role_envs {
+            env.extend(role_envs.iter().map(|(k, v)| {
+                (
+                    k,
+                    EnvVar {
+                        name: k.clone(),
+                        value: Some(v.clone()),
+                        ..Default::default()
+                    },
+                )
+            }))
+        }
+
+        env.into_values().collect()
+    }
+
     pub fn validated_role_config(
         &self,
         resolved_product_image: &ResolvedProductImage,
@@ -880,15 +948,17 @@ fn resources_to_driver_props(
         ..
     } = &driver_config.resources
     {
-        let min_cores = cores_from_quantity(min.0.clone())?;
-        let max_cores = cores_from_quantity(max.0.clone())?;
-        // will have default value from resources to apply if nothing set specifically
-        props.insert("spark.driver.cores".to_string(), max_cores.clone());
+        let driver_cores = cores_from_quantity(max.0.clone())?;
+        // take rounded value for driver.cores but actual values for the pod
+        props.insert("spark.driver.cores".to_string(), driver_cores.clone());
         props.insert(
             "spark.kubernetes.driver.request.cores".to_string(),
-            min_cores,
+            min.0.clone(),
         );
-        props.insert("spark.kubernetes.driver.limit.cores".to_string(), max_cores);
+        props.insert(
+            "spark.kubernetes.driver.limit.cores".to_string(),
+            max.0.clone(),
+        );
     }
 
     if let Resources {
@@ -920,17 +990,16 @@ fn resources_to_executor_props(
         ..
     } = &executor_config.resources
     {
-        let min_cores = cores_from_quantity(min.0.clone())?;
-        let max_cores = cores_from_quantity(max.0.clone())?;
-        // will have default value from resources to apply if nothing set specifically
-        props.insert("spark.executor.cores".to_string(), max_cores.clone());
+        let executor_cores = cores_from_quantity(max.0.clone())?;
+        // take rounded value for executor.cores (to determine the parallelism) but actual values for the pod
+        props.insert("spark.executor.cores".to_string(), executor_cores.clone());
         props.insert(
             "spark.kubernetes.executor.request.cores".to_string(),
-            min_cores,
+            min.0.clone(),
         );
         props.insert(
             "spark.kubernetes.executor.limit.cores".to_string(),
-            max_cores,
+            max.0.clone(),
         );
     }
 
@@ -950,6 +1019,9 @@ fn resources_to_executor_props(
 
 #[cfg(test)]
 mod tests {
+
+    use super::*;
+
     use crate::{cores_from_quantity, resources_to_executor_props, RoleConfig};
     use crate::{resources_to_driver_props, SparkApplication};
     use crate::{Quantity, SparkStorageConfig};
@@ -1099,7 +1171,7 @@ mod tests {
                 enable_vector_agent: false,
                 containers: BTreeMap::new(),
             },
-            volume_mounts: None,
+            volume_mounts: Default::default(),
             affinity: StackableAffinity::default(),
         };
 
@@ -1116,7 +1188,7 @@ mod tests {
             ),
             (
                 "spark.kubernetes.driver.request.cores".to_string(),
-                "1".to_string(),
+                "250m".to_string(),
             ),
         ]
         .into_iter()
@@ -1143,7 +1215,7 @@ mod tests {
                 enable_vector_agent: false,
                 containers: BTreeMap::new(),
             },
-            volume_mounts: None,
+            volume_mounts: Default::default(),
             affinity: StackableAffinity::default(),
         };
 
@@ -1156,7 +1228,7 @@ mod tests {
             ("spark.executor.memory".to_string(), "128m".to_string()), // 128 and not 512 because memory overhead is subtracted
             (
                 "spark.kubernetes.executor.request.cores".to_string(),
-                "1".to_string(),
+                "250m".to_string(),
             ),
             (
                 "spark.kubernetes.executor.limit.cores".to_string(),
@@ -1228,5 +1300,56 @@ mod tests {
         .collect();
 
         assert_eq!(expected, validated_config);
+    }
+
+    #[test]
+    fn test_job_volume_mounts() {
+        let spark_application = serde_yaml::from_str::<SparkApplication>(indoc! {r#"
+            ---
+            apiVersion: spark.stackable.tech/v1alpha1
+            kind: SparkApplication
+            metadata:
+              name: spark-examples
+            spec:
+              mode: cluster
+              mainApplicationFile: test.py
+              sparkImage:
+                productVersion: 1.2.3
+              job:
+                config:
+                  volumeMounts:
+                    - name: keytab
+                      mountPath: /kerberos
+              volumes:
+                - name: keytab
+                  configMap:
+                    name: keytab
+        "#})
+        .unwrap();
+
+        let got = spark_application.spark_job_volume_mounts(&None, &None);
+
+        let expected = vec![
+            VolumeMount {
+                mount_path: "/stackable/spark/driver-pod-templates".into(),
+                mount_propagation: None,
+                name: "driver-pod-template".into(),
+                ..VolumeMount::default()
+            },
+            VolumeMount {
+                mount_path: "/stackable/spark/executor-pod-templates".into(),
+                mount_propagation: None,
+                name: "executor-pod-template".into(),
+                ..VolumeMount::default()
+            },
+            VolumeMount {
+                mount_path: "/kerberos".into(),
+                mount_propagation: None,
+                name: "keytab".into(),
+                ..VolumeMount::default()
+            },
+        ];
+
+        assert_eq!(got, expected);
     }
 }