+++ rust/operator-binary/src/history/history_controller.rs
@@ -4,15 +4,18 @@ use crate::Ctx;
 use product_config::{types::PropertyNameKind, writer::to_java_properties_string};
 use stackable_operator::kvp::{Label, Labels, ObjectLabels};
 use stackable_operator::{
-    builder::{ConfigMapBuilder, ContainerBuilder, ObjectMetaBuilder, PodBuilder, VolumeBuilder},
+    builder::{
+        configmap::ConfigMapBuilder,
+        meta::ObjectMetaBuilder,
+        pod::{container::ContainerBuilder, volume::VolumeBuilder, PodBuilder},
+    },
     cluster_resources::{ClusterResourceApplyStrategy, ClusterResources},
     commons::product_image_selection::ResolvedProductImage,
     k8s_openapi::{
         api::{
             apps::v1::{StatefulSet, StatefulSetSpec},
             core::v1::{
-                ConfigMap, EnvVar, PodSecurityContext, Service, ServiceAccount, ServicePort,
-                ServiceSpec,
+                ConfigMap, PodSecurityContext, Service, ServiceAccount, ServicePort, ServiceSpec,
             },
             rbac::v1::{ClusterRole, RoleBinding, RoleRef, Subject},
         },
@@ -32,14 +35,15 @@ use stackable_operator::{
     role_utils::RoleGroupRef,
     time::Duration,
 };
+use stackable_spark_k8s_crd::constants::METRICS_PORT;
 use stackable_spark_k8s_crd::{
     constants::{
         ACCESS_KEY_ID, APP_NAME, HISTORY_CONTROLLER_NAME, HISTORY_ROLE_NAME,
-        JVM_SECURITY_PROPERTIES_FILE, LOG4J2_CONFIG_FILE, MAX_SPARK_LOG_FILES_SIZE, OPERATOR_NAME,
-        SECRET_ACCESS_KEY, SPARK_CLUSTER_ROLE, SPARK_DEFAULTS_FILE_NAME, SPARK_IMAGE_BASE_NAME,
-        SPARK_UID, STACKABLE_TLS_STORE_PASSWORD, STACKABLE_TRUST_STORE, VOLUME_MOUNT_NAME_CONFIG,
-        VOLUME_MOUNT_NAME_LOG, VOLUME_MOUNT_NAME_LOG_CONFIG, VOLUME_MOUNT_PATH_CONFIG,
-        VOLUME_MOUNT_PATH_LOG, VOLUME_MOUNT_PATH_LOG_CONFIG,
+        JVM_SECURITY_PROPERTIES_FILE, MAX_SPARK_LOG_FILES_SIZE, OPERATOR_NAME, SECRET_ACCESS_KEY,
+        SPARK_CLUSTER_ROLE, SPARK_DEFAULTS_FILE_NAME, SPARK_IMAGE_BASE_NAME, SPARK_UID,
+        STACKABLE_TRUST_STORE, VOLUME_MOUNT_NAME_CONFIG, VOLUME_MOUNT_NAME_LOG,
+        VOLUME_MOUNT_NAME_LOG_CONFIG, VOLUME_MOUNT_PATH_CONFIG, VOLUME_MOUNT_PATH_LOG,
+        VOLUME_MOUNT_PATH_LOG_CONFIG,
     },
     history,
     history::{HistoryConfig, SparkHistoryServer, SparkHistoryServerContainer},
@@ -50,12 +54,10 @@ use std::collections::HashMap;
 use std::{collections::BTreeMap, sync::Arc};
 
 use snafu::{OptionExt, ResultExt, Snafu};
-use stackable_operator::builder::resources::ResourceRequirementsBuilder;
+use stackable_operator::builder::pod::resources::ResourceRequirementsBuilder;
 use stackable_operator::k8s_openapi::DeepMerge;
 use stackable_operator::logging::controller::ReconcilerError;
 use strum::{EnumDiscriminants, IntoStaticStr};
-
-const METRICS_PORT: u16 = 18081;
 
 #[derive(Snafu, Debug, EnumDiscriminants)]
 #[strum_discriminants(derive(IntoStaticStr))]
@@ -65,36 +67,36 @@ pub enum Error {
     ObjectHasNoNamespace,
     #[snafu(display("invalid config map {name}"))]
     InvalidConfigMap {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::builder::configmap::Error,
         name: String,
     },
     #[snafu(display("invalid history container name"))]
     InvalidContainerName {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::builder::pod::container::Error,
     },
     #[snafu(display("object is missing metadata to build owner reference"))]
     ObjectMissingMetadataForOwnerRef {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::builder::meta::Error,
     },
     #[snafu(display("failed to update the history server deployment"))]
     ApplyDeployment {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::cluster_resources::Error,
     },
     #[snafu(display("failed to update history server config map"))]
     ApplyConfigMap {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::cluster_resources::Error,
     },
     #[snafu(display("failed to update history server service"))]
     ApplyService {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::cluster_resources::Error,
     },
     #[snafu(display("failed to apply role ServiceAccount"))]
     ApplyServiceAccount {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::cluster_resources::Error,
     },
     #[snafu(display("failed to apply global RoleBinding"))]
     ApplyRoleBinding {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::cluster_resources::Error,
     },
     #[snafu(display("product config validation failed"))]
     ProductConfigValidation {
@@ -114,11 +116,11 @@ pub enum Error {
     },
     #[snafu(display("failed to create cluster resources"))]
     CreateClusterResources {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::cluster_resources::Error,
     },
     #[snafu(display("failed to delete orphaned resources"))]
     DeleteOrphanedResources {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::cluster_resources::Error,
     },
     #[snafu(display("failed to resolve the Vector aggregator address"))]
     ResolveVectorAggregatorAddress { source: product_logging::Error },
@@ -149,7 +151,7 @@ pub enum Error {
 
     #[snafu(display("failed to build Metadata"))]
     MetadataBuild {
-        source: stackable_operator::builder::ObjectMetaBuilderError,
+        source: stackable_operator::builder::meta::Error,
     },
 
     #[snafu(display("failed to get required Labels"))]
@@ -189,7 +191,7 @@ pub async fn reconcile(shs: Arc<SparkHistoryServer>, ctx: Arc<Ctx>) -> Result<Ac
     let resolved_product_image = shs
         .spec
         .image
-        .resolve(SPARK_IMAGE_BASE_NAME, crate::built_info::CARGO_PKG_VERSION);
+        .resolve(SPARK_IMAGE_BASE_NAME, crate::built_info::PKG_VERSION);
     let s3_log_dir = S3LogDir::resolve(
         Some(&shs.spec.log_file_directory),
         shs.metadata.namespace.clone(),
@@ -310,6 +312,7 @@ pub fn error_policy(_obj: Arc<SparkHistoryServer>, _error: &Error, _ctx: Arc<Ctx
     Action::requeue(*Duration::from_secs(5))
 }
 
+#[allow(clippy::result_large_err)]
 fn build_config_map(
     shs: &SparkHistoryServer,
     config: &HashMap<PropertyNameKind, BTreeMap<String, String>>,
@@ -371,6 +374,7 @@ fn build_config_map(
         .context(InvalidConfigMapSnafu { name: cm_name })
 }
 
+#[allow(clippy::result_large_err)]
 fn build_stateful_set(
     shs: &SparkHistoryServer,
     resolved_product_image: &ResolvedProductImage,
@@ -434,6 +438,12 @@ fn build_stateful_set(
             ..PodSecurityContext::default()
         });
 
+    let role_group = shs
+        .rolegroup(rolegroupref)
+        .with_context(|_| CannotRetrieveRoleGroupSnafu)?;
+
+    let merged_env = shs.merged_env(s3_log_dir, role_group.config.env_overrides);
+
     let container_name = "spark-history";
     let container = ContainerBuilder::new(container_name)
         .context(InvalidContainerNameSnafu)?
@@ -443,7 +453,7 @@ fn build_stateful_set(
         .args(command_args(s3_log_dir))
         .add_container_port("http", 18080)
         .add_container_port("metrics", METRICS_PORT.into())
-        .add_env_vars(env_vars(s3_log_dir))
+        .add_env_vars(merged_env)
         .add_volume_mounts(s3_log_dir.volume_mounts())
         .add_volume_mount(VOLUME_MOUNT_NAME_CONFIG, VOLUME_MOUNT_PATH_CONFIG)
         .add_volume_mount(VOLUME_MOUNT_NAME_LOG_CONFIG, VOLUME_MOUNT_PATH_LOG_CONFIG)
@@ -471,10 +481,6 @@ fn build_stateful_set(
 
     let mut pod_template = pb.build_template();
     pod_template.merge_from(shs.role().config.pod_overrides.clone());
-    let role_group = shs
-        .rolegroup(rolegroupref)
-        .with_context(|_| CannotRetrieveRoleGroupSnafu)?;
-
     pod_template.merge_from(role_group.config.pod_overrides);
 
     Ok(StatefulSet {
@@ -512,6 +518,7 @@ fn build_stateful_set(
     })
 }
 
+#[allow(clippy::result_large_err)]
 fn build_service(
     shs: &SparkHistoryServer,
     app_version_label: &str,
@@ -577,6 +584,7 @@ fn build_service(
     })
 }
 
+#[allow(clippy::result_large_err)]
 fn build_history_role_serviceaccount(
     shs: &SparkHistoryServer,
     app_version_label: &str,
@@ -617,6 +625,7 @@ fn build_history_role_serviceaccount(
     Ok((sa, binding))
 }
 
+#[allow(clippy::result_large_err)]
 fn spark_defaults(
     shs: &SparkHistoryServer,
     s3_log_dir: &S3LogDir,
@@ -661,46 +670,6 @@ fn command_args(s3logdir: &S3LogDir) -> Vec<String> {
     vec![String::from("-c"), command.join(" && ")]
 }
 
-fn env_vars(s3logdir: &S3LogDir) -> Vec<EnvVar> {
-    let mut vars: Vec<EnvVar> = vec![];
-
-    // This env var prevents the history server from detaching itself from the
-    // start script because this leads to the Pod terminating immediately.
-    vars.push(EnvVar {
-        name: "SPARK_NO_DAEMONIZE".to_string(),
-        value: Some("true".into()),
-        value_from: None,
-    });
-    vars.push(EnvVar {
-        name: "SPARK_DAEMON_CLASSPATH".to_string(),
-        value: Some("/stackable/spark/extra-jars/*".into()),
-        value_from: None,
-    });
-
-    let mut history_opts = vec![
-        format!("-Dlog4j.configurationFile={VOLUME_MOUNT_PATH_LOG_CONFIG}/{LOG4J2_CONFIG_FILE}"),
-        format!(
-            "-Djava.security.properties={VOLUME_MOUNT_PATH_CONFIG}/{JVM_SECURITY_PROPERTIES_FILE}"
-        ),
-        format!("-javaagent:/stackable/jmx/jmx_prometheus_javaagent.jar={METRICS_PORT}:/stackable/jmx/config.yaml")
-    ];
-    if tlscerts::tls_secret_name(&s3logdir.bucket.connection).is_some() {
-        history_opts.extend(vec![
-            format!("-Djavax.net.ssl.trustStore={STACKABLE_TRUST_STORE}/truststore.p12"),
-            format!("-Djavax.net.ssl.trustStorePassword={STACKABLE_TLS_STORE_PASSWORD}"),
-            format!("-Djavax.net.ssl.trustStoreType=pkcs12"),
-        ]);
-    }
-
-    vars.push(EnvVar {
-        name: "SPARK_HISTORY_OPTS".to_string(),
-        value: Some(history_opts.join(" ")),
-        value_from: None,
-    });
-    // if TLS is enabled build truststore
-    vars
-}
-
 fn labels<'a, T>(
     shs: &'a T,
     app_version_label: &'a str,
@@ -720,6 +689,7 @@ fn labels<'a, T>(
 /// Return the Spark properties for the cleaner role group (if any).
 /// There should be only one role group with "cleaner=true" and this
 /// group should have a replica count of 0 or 1.
+#[allow(clippy::result_large_err)]
 fn cleaner_config(
     shs: &SparkHistoryServer,
     rolegroup_ref: &RoleGroupRef<SparkHistoryServer>,