+++ rust/operator-binary/src/history/operations/pdb.rs
@@ -12,12 +12,12 @@ use stackable_spark_k8s_crd::{
 pub enum Error {
     #[snafu(display("Cannot create PodDisruptionBudget for role [{role}]"))]
     CreatePdb {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::builder::pdb::Error,
         role: String,
     },
     #[snafu(display("Cannot apply PodDisruptionBudget [{name}]"))]
     ApplyPdb {
-        source: stackable_operator::error::Error,
+        source: stackable_operator::cluster_resources::Error,
         name: String,
     },
 }