+++ rust/operator-binary/src/main.rs
@@ -26,8 +26,6 @@ use tracing_futures::Instrument;
 
 mod built_info {
     include!(concat!(env!("OUT_DIR"), "/built.rs"));
-    pub const TARGET_PLATFORM: Option<&str> = option_env!("TARGET");
-    pub const CARGO_PKG_VERSION: &str = env!("CARGO_PKG_VERSION");
 }
 
 #[derive(Parser)]
@@ -51,8 +49,8 @@ async fn main() -> anyhow::Result<()> {
     let opts = Opts::parse();
     match opts.cmd {
         Command::Crd => {
-            SparkApplication::print_yaml_schema(built_info::CARGO_PKG_VERSION)?;
-            SparkHistoryServer::print_yaml_schema(built_info::CARGO_PKG_VERSION)?;
+            SparkApplication::print_yaml_schema(built_info::PKG_VERSION)?;
+            SparkHistoryServer::print_yaml_schema(built_info::PKG_VERSION)?;
         }
         Command::Run(ProductOperatorRun {
             product_config,
@@ -68,7 +66,7 @@ async fn main() -> anyhow::Result<()> {
                 crate_description!(),
                 crate_version!(),
                 built_info::GIT_VERSION,
-                built_info::TARGET_PLATFORM.unwrap_or("unknown target"),
+                built_info::TARGET,
                 built_info::BUILT_TIME_UTC,
                 built_info::RUSTC_VERSION,
             );